rename(version_utils): make module version_utils private
